+++ boioot/apps/backend/src/Boioot.Application/Features/Plans/DTOs/SetPlanLimitRequest.cs
@@ -2,6 +2,6 @@ namespace Boioot.Application.Features.Plans.DTOs;
 
 public class SetPlanLimitRequest
 {
-    /// <summary>The new limit value. Use -1 for unlimited.</summary>
-    public decimal Value { get; set; }
+    /// <summary>The new limit value. Must be an integer. Use -1 for unlimited.</summary>
+    public int Value { get; set; }
 }
