+++ boioot/apps/backend/src/Boioot.Application/Features/Plans/Interfaces/IAdminPlanService.cs
@@ -18,8 +18,8 @@ public interface IAdminPlanService
     /// <summary>Creates a full copy of the plan (limits + features). New plan starts as inactive/private.</summary>
     Task<PlanDetailResponse> DuplicatePlanAsync(Guid sourcePlanId, CancellationToken ct = default);
 
-    /// <summary>Set or update the value for a named limit on a plan. Creates the row if missing.</summary>
-    Task<PlanLimitItem> SetLimitAsync(Guid planId, string limitKey, decimal value, CancellationToken ct = default);
+    /// <summary>Set or update the value for a named limit on a plan. Creates the row if missing. Value must be integer (-1 = unlimited).</summary>
+    Task<PlanLimitItem> SetLimitAsync(Guid planId, string limitKey, int value, CancellationToken ct = default);
 
     /// <summary>Enable or disable a named feature on a plan. Creates the row if missing.</summary>
     Task<PlanFeatureItem> SetFeatureAsync(Guid planId, string featureKey, bool isEnabled, CancellationToken ct = default);
